--- components/location.js
@@ -1,11 +1,0 @@
-import React from 'react'
-import { View } from 'react-native';
-
-export default class LocationPick extends React.Component {
-
-    render() {
-        return (
-            <View></View>
-        );
-    }
-}
